LPS-48728 don't expose implementation details
The `deleteTrashEntry` methods are implementation details of Liferay's
trash capability. Don't expose them in the interface and register the
events in the capability itself.
