+++ portal-impl/src/com/liferay/portal/repository/liferayrepository/LiferayRepositoryRegistryPlugin.java
@@ -14,12 +14,7 @@
 
 package com.liferay.portal.repository.liferayrepository;
 
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.repository.capabilities.TrashCapability;
-import com.liferay.portal.kernel.repository.event.RepositoryEventListener;
-import com.liferay.portal.kernel.repository.event.RepositoryEventType;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.repository.registry.BaseRepositoryRegistryPlugin;
 import com.liferay.portal.kernel.repository.registry.CapabilityRegistry;
 import com.liferay.portal.kernel.repository.registry.RepositoryCreator;
@@ -55,41 +50,16 @@ public class LiferayRepositoryRegistryPlugin
 	public void registerRepositoryEventListeners(
 		RepositoryEventRegistry repositoryEventRegistry) {
 
-		repositoryEventRegistry.registerRepositoryEventListener(
-			RepositoryEventType.Delete.class, FileEntry.class,
-			new DeleteFileEntryRepositoryEventListener());
-		repositoryEventRegistry.registerRepositoryEventListener(
-			RepositoryEventType.Delete.class, Folder.class,
-			new DeleteFolderRepositoryEventListener());
+		_liferayTrashCapability.registerRepositoryEventListeners(
+			repositoryEventRegistry);
 	}
 
 	public void setRepositoryCreator(RepositoryCreator repositoryCreator) {
 		_repositoryCreator = repositoryCreator;
 	}
 
-	private TrashCapability _liferayTrashCapability =
+	private LiferayTrashCapability _liferayTrashCapability =
 		new LiferayTrashCapability();
 	private RepositoryCreator _repositoryCreator;
-
-	private class DeleteFileEntryRepositoryEventListener
-		implements RepositoryEventListener
-			<RepositoryEventType.Delete, FileEntry> {
-
-		@Override
-		public void execute(FileEntry fileEntry) throws PortalException {
-			_liferayTrashCapability.deleteTrashEntry(fileEntry);
-		}
-
-	}
-
-	private class DeleteFolderRepositoryEventListener
-		implements RepositoryEventListener<RepositoryEventType.Delete, Folder> {
-
-		@Override
-		public void execute(Folder folder) throws PortalException {
-			_liferayTrashCapability.deleteTrashEntry(folder);
-		}
-
-	}
 
 }
