+++ portal-service/src/com/liferay/portal/kernel/repository/capabilities/TrashCapability.java
@@ -30,10 +30,6 @@ public interface TrashCapability extends Capability {
 
 	public void deleteTrashEntries(long repositoryId) throws PortalException;
 
-	public void deleteTrashEntry(FileEntry fileEntry) throws PortalException;
-
-	public void deleteTrashEntry(Folder folder) throws PortalException;
-
 	public boolean isInTrash(Folder folder) throws PortalException;
 
 	public FileEntry moveFileEntryFromTrash(
